Add server folder
