+++ src/pages/[spaceId]/[slug]/index.tsx
@@ -2,11 +2,11 @@ import { CHAT_PER_PAGE } from '@/constants/chat'
 import ChatPage, { ChatPageProps } from '@/modules/chat/ChatPage'
 import { getPostsFromCache } from '@/pages/api/posts'
 import { AppCommonProps } from '@/pages/_app'
+import { prefetchBlockedEntities } from '@/server/moderation'
 import { getPostQuery } from '@/services/api/query'
 import { getCommentIdsQueryKey } from '@/services/subsocial/commentIds'
 import { getSubsocialApi } from '@/subsocial-query/subsocial/connection'
 import { getCommonStaticProps } from '@/utils/page'
-import { prefetchBlockedEntities } from '@/utils/server'
 import { getIdFromSlug } from '@/utils/slug'
 import { validateNumber } from '@/utils/strings'
 import { dehydrate, QueryClient } from '@tanstack/react-query'
