+++ src/pages/[spaceId]/index.tsx
@@ -3,6 +3,7 @@ import {
   getSpaceIdFromAlias,
 } from '@/constants/chat-room'
 import HomePage, { HomePageProps } from '@/modules/chat/HomePage'
+import { prefetchBlockedEntities } from '@/server/moderation'
 import { getPostQuery } from '@/services/api/query'
 import { getCommentIdsQueryKey } from '@/services/subsocial/commentIds'
 import { getChatIdsBySpaceIdQuery } from '@/services/subsocial/posts'
@@ -10,7 +11,6 @@ import { getSpaceBySpaceIdQuery } from '@/services/subsocial/spaces'
 import { getSubsocialApi } from '@/subsocial-query/subsocial/connection'
 import { getMainSpaceId } from '@/utils/env/client'
 import { getCommonStaticProps } from '@/utils/page'
-import { prefetchBlockedEntities } from '@/utils/server'
 import { validateNumber } from '@/utils/strings'
 import { PostData } from '@subsocial/api/types'
 import { dehydrate, QueryClient } from '@tanstack/react-query'
